backend/agents/heal_agent.py
"""
RIFT 2026 — Heal Agent

Generates targeted fixes for each classified error, creates the fix branch,
and commits changes with [AI-AGENT] prefix.
"""
import logging
import re
from pathlib import Path

from git import Repo

from models import BugType, ErrorInfo, Fix, FixStatus
from services.git_service import git_service
from utils import format_branch_name, format_commit_message

logger = logging.getLogger("rift.heal_agent")


class HealAgent:
    """Agent that applies code fixes and commits them."""

    def run(
        self,
        repo_path: str,
        errors: list[ErrorInfo],
        team_name: str,
        leader_name: str,
        iteration: int,
    ) -> tuple[list[Fix], str, int]:
        """
        Apply fixes for all errors, create branch (on first iteration),
        commit each fix.

        Args:
            repo_path: Path to cloned repo.
            errors: Classified errors from AnalyzeAgent.
            team_name: Team name.
            leader_name: Leader name.
            iteration: Current iteration number.

        Returns:
            Tuple of (list[Fix], branch_name, total_new_commits).
        """
        repo = git_service.get_repo(repo_path)
        branch_name = format_branch_name(team_name, leader_name)

        # Create branch on first iteration
        if iteration == 1:
            git_service.create_branch(repo, team_name, leader_name)
        else:
            # Ensure we're on the right branch
            if repo.active_branch.name != branch_name:
                repo.git.checkout(branch_name)

        fixes: list[Fix] = []
        commit_count = 0

        for error in errors:
            fix = self._generate_fix(repo_path, error)
            if fix is None:
                logger.warning(
                    f"Could not generate fix for {error.file}:{error.line_number}"
                )
                continue

            # Apply the fix to the file
            applied = self._apply_fix(repo_path, error, fix)
            if not applied:
                fix.status = FixStatus.FAILED
                fixes.append(fix)
                continue

            # Commit
            try:
                commit_msg = format_commit_message(
                    error.bug_type.value, error.file, error.line_number
                )
                fix.commit_message = commit_msg
                git_service.commit_fix(repo, error.file, commit_msg)
                fix.status = FixStatus.APPLIED
                commit_count += 1
            except Exception as e:
                logger.error(f"Commit failed for {error.file}: {e}")
                fix.status = FixStatus.FAILED

            fixes.append(fix)

        # Push all commits at once
        if commit_count > 0:
            try:
                git_service.push(repo, branch_name)
            except Exception as e:
                logger.error(f"Push failed: {e}")

        logger.info(
            f"Heal iteration {iteration}: {commit_count} fixes applied, "
            f"{len(fixes)} total"
        )

        return fixes, branch_name, commit_count

    # ---- Fix generators per error type ----

    def _generate_fix(self, repo_path: str, error: ErrorInfo) -> Fix | None:
        """Generate a Fix object for the given error."""
        file_path = Path(repo_path) / error.file
        if not file_path.exists():
            return None

        try:
            content = file_path.read_text(encoding="utf-8", errors="replace")
            lines = content.splitlines(keepends=True)
        except Exception:
            return None

        if error.line_number < 1 or error.line_number > len(lines):
            return None

        original_line = lines[error.line_number - 1]

        fix_generators = {
            BugType.IMPORT: self._fix_import,
            BugType.INDENTATION: self._fix_indentation,
            BugType.SYNTAX: self._fix_syntax,
            BugType.TYPE_ERROR: self._fix_type_error,
            BugType.LINTING: self._fix_linting,
            BugType.LOGIC: self._fix_logic,
        }

        generator = fix_generators.get(error.bug_type, self._fix_generic)
        fixed_line = generator(error, lines, error.line_number)

        if fixed_line is None:
            return None

        return Fix(
            file=error.file,
            bug_type=error.bug_type,
            line_number=error.line_number,
            original_code=original_line.rstrip(),
            fixed_code=fixed_line.rstrip(),
            status=FixStatus.PENDING,
        )

    def _fix_import(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix import errors — add missing import or fix module name."""
        msg = error.message.lower()

        # Extract module name from "No module named 'xyz'"
        match = re.search(r"no module named\s+['\"](\S+)['\"]", msg)
        if match:
            module = match.group(1)
            # Common typo fixes
            typo_fixes = {
                "colections": "collections",
                "ittertools": "itertools",
                "jsons": "json",
                "maths": "math",
                "os.paths": "os.path",
                "sys.path": "sys",
                "requets": "requests",
                "numppy": "numpy",
                "pands": "pandas",
            }
            if module in typo_fixes:
                return lines[line_num - 1].replace(module, typo_fixes[module])

        # "cannot import name 'X' from 'Y'"
        match = re.search(r"cannot import name\s+['\"](\S+)['\"]", msg)
        if match:
            name = match.group(1)
            # Try to comment out the bad import and add a TODO
            return f"# [AI-AGENT] Removed broken import: {name}\n"

        # Generic: comment out the broken import
        return f"# [AI-AGENT] Fixed: {lines[line_num - 1].rstrip()}\n"

    def _fix_indentation(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix indentation — align with context."""
        if line_num < 2:
            return lines[line_num - 1].lstrip()

        # Look at previous non-empty line for reference indentation
        prev_indent = ""
        for i in range(line_num - 2, max(-1, line_num - 6), -1):
            prev_line = lines[i]
            if prev_line.strip():
                prev_indent = re.match(r"^(\s*)", prev_line).group(1)
                # If previous line ends with ':', add one indent level
                if prev_line.rstrip().endswith(":"):
                    prev_indent += "    "
                break

        current_content = lines[line_num - 1].lstrip()
        # Dedent keywords
        if current_content.startswith(("else:", "elif ", "except", "finally:", "except:")):
            # Reduce one indent level
            if len(prev_indent) >= 4:
                prev_indent = prev_indent[:-4]

        return prev_indent + current_content

    def _fix_syntax(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix syntax errors — common patterns."""
        line = lines[line_num - 1]

        # Missing colon at end of def/class/if/for/while/with/else/elif/try/except
        stripped = line.rstrip()
        keywords = ("def ", "class ", "if ", "for ", "while ", "with ", "else", "elif ", "try", "except", "finally")
        for kw in keywords:
            if stripped.lstrip().startswith(kw) and not stripped.endswith(":"):
                # Remove trailing content that shouldn't be there
                return stripped + ":\n"

        # Unmatched parentheses — add missing closing paren
        open_count = line.count("(") - line.count(")")
        if open_count > 0:
            return line.rstrip() + ")" * open_count + "\n"

        open_count = line.count("[") - line.count("]")
        if open_count > 0:
            return line.rstrip() + "]" * open_count + "\n"

        # Missing quotes
        single_quotes = line.count("'")
        double_quotes = line.count('"')
        if single_quotes % 2 != 0:
            return line.rstrip() + "'\n"
        if double_quotes % 2 != 0:
            return line.rstrip() + '"\n'

        return line  # Return unchanged if no pattern matched

    def _fix_type_error(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix type errors — add type conversions."""
        msg = error.message.lower()
        line = lines[line_num - 1]

        # "unsupported operand type(s) for +: 'int' and 'str'"
        if "str" in msg and "int" in msg:
            # Wrap potential string concatenation with str()
            line = re.sub(
                r'(\+\s*)(\w+)',
                lambda m: m.group(1) + f"str({m.group(2)})",
                line,
                count=1,
            )
            return line

        # "'NoneType' object is not subscriptable/iterable"
        if "nonetype" in msg:
            indent = re.match(r"^(\s*)", line).group(1)
            var_match = re.search(r"(\w+)\s*[\[\(.]", line)
            if var_match:
                var = var_match.group(1)
                return f"{indent}if {var} is not None:\n{indent}    {line.lstrip()}"

        return line

    def _fix_linting(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix common linting issues."""
        line = lines[line_num - 1]
        msg = error.message.lower()

        # Trailing whitespace
        if "trailing whitespace" in msg or "W291" in error.message:
            return line.rstrip() + "\n"

        # Missing newline at end of file
        if "no newline" in msg or "W292" in error.message:
            return line.rstrip() + "\n"

        # Unused import — comment it out
        if "unused" in msg and "import" in msg:
            return f"# {line.rstrip()}  # noqa: removed unused import\n"

        # Line too long
        if "line too long" in msg or "E501" in error.message:
            return line  # Keep as-is, hard to auto-fix

        return line.rstrip() + "\n"

    def _fix_logic(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Fix logic errors — attempt assertion fixes."""
        line = lines[line_num - 1]
        msg = error.message

        # If assertion shows expected vs actual, try to fix the function
        # This is limited — we add a comment for human review
        match = re.search(r"assert\s+(.+?)\s*==\s*(.+)", line)
        if match:
            # Can't reliably fix logic, add a marker
            return line  # Keep assertion, fix should be in the function

        return line

    def _fix_generic(
        self, error: ErrorInfo, lines: list[str], line_num: int
    ) -> str | None:
        """Generic fallback fix."""
        return lines[line_num - 1]

    def _apply_fix(
        self, repo_path: str, error: ErrorInfo, fix: Fix
    ) -> bool:
        """Write the fix to the file."""
        file_path = Path(repo_path) / error.file
        try:
            content = file_path.read_text(encoding="utf-8", errors="replace")
            lines = content.splitlines(keepends=True)

            if error.line_number < 1 or error.line_number > len(lines):
                return False

            # Replace the line
            lines[error.line_number - 1] = fix.fixed_code + "\n"

            file_path.write_text("".join(lines), encoding="utf-8")
            return True
        except Exception as e:
            logger.error(f"Failed to apply fix to {error.file}: {e}")
            return False


# Singleton
heal_agent = HealAgent()
