backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
gitpython==3.1.43
docker==7.1.0
httpx==0.27.2
python-dotenv==1.0.1
crewai>=1.9.0
crewai-tools>=0.17.0
litellm>=1.40.0
anthropic>=0.30.0
google-generativeai>=0.7.0
